Create a new io package and move the FastQInputFormat and associated classes into it. We do this to avoid circular dependencies. Many MR jobs will probably want to use this input format and we don't want those jobs to have to depend on the correct package.

+++ src/main/java/contrail/correct/FastQToAvro.java
@@ -1,21 +1,15 @@
 package contrail.correct;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.avro.mapred.AvroCollector;
-import org.apache.avro.mapred.AvroJob;
-import org.apache.avro.mapred.AvroMapper;
 import org.apache.avro.mapred.AvroWrapper;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.filecache.DistributedCache;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.NullWritable;
-import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.FileInputFormat;
 import org.apache.hadoop.mapred.FileOutputFormat;
 import org.apache.hadoop.mapred.JobClient;
@@ -33,12 +27,11 @@ import contrail.stages.BuildGraphAvro;
 import contrail.stages.ContrailParameters;
 import contrail.stages.ParameterDefinition;
 import contrail.stages.Stage;
-import java.util.*;
-import contrail.correct.FastQText;
-import contrail.correct.FastQInputFormat;
+import contrail.io.FastQText;
+import contrail.io.FastQInputFormat;
 
 
-/** MapReduce job to convert a FastQ File to Avro. 
+/** MapReduce job to convert a FastQ File to Avro.
  * Uses FastQInputFormat.
  * @author dnettem
  */
@@ -47,7 +40,7 @@ public class FastQToAvro extends Stage {
 /*  private static*/ final Logger sLogger = Logger.getLogger(BuildGraphAvro.class);
 
   protected Map<String, ParameterDefinition> createParameterDefinitions() {
-    HashMap<String, ParameterDefinition> defs = new HashMap<String, 
+    HashMap<String, ParameterDefinition> defs = new HashMap<String,
         ParameterDefinition>();
 
     defs.putAll(super.createParameterDefinitions());
@@ -58,73 +51,73 @@ public class FastQToAvro extends Stage {
     return Collections.unmodifiableMap(defs);
   }
 
-  public static class FastQToAvroMapper extends MapReduceBase 
+  public static class FastQToAvroMapper extends MapReduceBase
   implements Mapper<LongWritable, FastQText, AvroWrapper<FastQRecord>, NullWritable>{
 
     private FastQRecord read = new FastQRecord();
     private AvroWrapper<FastQRecord> out_wrapper = new AvroWrapper<FastQRecord>(read);
-    
+
     public void configure(JobConf job)
     {
       Configuration conf = new Configuration();
       FileSystem hdfs;
     }
-    
+
     public void map(LongWritable line, FastQText record,
         OutputCollector<AvroWrapper<FastQRecord>, NullWritable> output, Reporter reporter)
             throws IOException {
-      
-      read.id = record.getId(); 
+
+      read.id = record.getId();
       read.read = record.getDna();
       read.qvalue = record.getQValue();
-      
-      output.collect(out_wrapper, NullWritable.get());     
+
+      output.collect(out_wrapper, NullWritable.get());
     }
   }
 
   public RunningJob runJob() throws Exception {
-    
+
     JobConf conf = new JobConf(FastQToAvro.class);
-    
+
     conf.setJobName("FastQToAvro");
     String inputPath, outputPath, datatype;
     Long splitSize;
     conf.setJobName("Rekey Data");
     String[] required_args = {"inputpath", "outputpath","splitSize"};
-    checkHasParametersOrDie(required_args);   
-    
+    checkHasParametersOrDie(required_args);
+
     inputPath = (String) stage_options.get("inputpath");
     outputPath = (String) stage_options.get("outputpath");
     splitSize = (Long) stage_options.get("splitSize");
-    
+
     //Sets the parameters in JobConf
     initializeJobConfiguration(conf);
     FileInputFormat.addInputPath(conf, new Path(inputPath));
     FileOutputFormat.setOutputPath(conf, new Path(outputPath));
-   
+
     // Input
     conf.setMapperClass(FastQToAvroMapper.class);
     conf.setInputFormat(FastQInputFormat.class);
     conf.setLong("FastQInputFormat.splitSize", splitSize);
     //Map Only Job
     conf.setNumReduceTasks(0);
-    
+
     // Delete the output directory if it exists already
     Path out_path = new Path(outputPath);
     if (FileSystem.get(conf).exists(out_path)) {
-      FileSystem.get(conf).delete(out_path, true);  
-    }     
-    
-    long starttime = System.currentTimeMillis();            
+      FileSystem.get(conf).delete(out_path, true);
+    }
+
+    long starttime = System.currentTimeMillis();
     RunningJob result = JobClient.runJob(conf);
     long endtime = System.currentTimeMillis();
-    float diff = (float) (((float) (endtime - starttime)) / 1000.0);
+    float diff = (float) ((endtime - starttime) / 1000.0);
     System.out.println("Runtime: " + diff + " s");
-    return result;   
+    return result;
   }
- 
+
   public static void main(String[] args) throws Exception {
-    
+
     int res = ToolRunner.run(new Configuration(), new FastQToAvro(), args);
     System.exit(res);
   }
